Split metric-chart.component, allow multiline while fetching server
metric-chart.component is split between metric-chart-static.component and metric-chart-dyn.component because they are quite different, used for different use cases, many mutually exclusive fields and having both together adds avoidable complexity

The "dynamic" component (ie. fetching data from server) now accepts several input metrics, which result in multiline chart.
