+++ src/hawkular-charts.ts
@@ -1,4 +1,5 @@
 // Public classes.
 export { HawkularChartsModule } from './modules/hawkular-charts.module';
-export { MetricChartComponent } from './components/metric-chart.component';
+import { MetricChartStaticComponent } from './components/metric-chart-static.component';
+import { MetricChartDynComponent } from './components/metric-chart-dyn.component';
 export { AvailChartComponent } from './components/avail-chart.component';
