+++ src/model/chart-options.ts
@@ -7,8 +7,8 @@ export class ChartOptions {
     public svg: any, // d3.Selection<any>
     public layout: ChartLayout,
     public axis: ComputedChartAxis,
-    public data: INumericDataPoint[],
-    public multiData: INamedMetric[],
+    public data?: INumericDataPoint[],
+    public multiData?: INamedMetric[],
     public tip?: any,
     public hideHighLowValues?: boolean,
     public interpolation?: string) {
